Migrate app dir owner check to SetupCheck API
